Return healthy flow unit from ThreadMetricsRca
